# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PrepSmart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List all recipes
      parameters:
        - in: query
          name: favoritesOnly
          schema:
            type: boolean
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a recipe by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRecipe
      tags: [recipes]
      summary: Update a recipe
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipeBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete a recipe
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /recipes/{id}/favorite:
    patch:
      operationId: toggleRecipeFavorite
      tags: [recipes]
      summary: Toggle recipe favorite status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Meal Plans
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Recipe:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        ingredients:
          type: array
          items:
            type: string
        instructions:
          type: string
        prepTime:
          type: ["integer", "null"]
        calories:
          type: ["integer", "null"]
        isFavorite:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - name
        - ingredients
        - instructions
        - prepTime
        - calories
        - isFavorite
        - createdAt
        - updatedAt
    CreateRecipeBody:
      type: object
      properties:
        name:
          type: string
        ingredients:
          type: array
          items:
            type: string
        instructions:
          type: string
        prepTime:
          type: ["integer", "null"]
        calories:
          type: ["integer", "null"]
      required:
        - name
        - ingredients
        - instructions
    UpdateRecipeBody:
      type: object
      properties:
        name:
          type: string
        ingredients:
          type: array
          items:
            type: string
        instructions:
          type: string
        prepTime:
          type: ["integer", "null"]
        calories:
          type: ["integer", "null"]

    # Meal Plans
...